% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTheme.R
\docType{data}
\name{BasicTheme}
\alias{BasicTheme}
\title{Add a simple ggplot theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 136.
}
\usage{
BasicTheme
}
\value{
Simple ggplot theme
}
\description{
Add a simple ggplot theme
}
\keyword{datasets}
