% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxToPolygon.R
\name{bboxToPolygon}
\alias{bboxToPolygon}
\title{Make a rectangular polygon from a st_*() object}
\usage{
bboxToPolygon(x)
}
\value{
Polygon
}
\description{
Make a rectangular polygon from a st_*() object
}
