% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Calculate coefficient of variation of a vector of numbers}
\usage{
cv(x, ...)
}
\value{
Coefficient of variation
}
\description{
Calculate coefficient of variation of a vector of numbers
}
