% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptyTheme.R
\docType{data}
\name{EmptyTheme}
\alias{EmptyTheme}
\title{Add an empty ggplot theme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 93.}
\usage{
EmptyTheme
}
\value{
Simple ggplot theme
}
\description{
Add a simple ggplot theme
}
\keyword{datasets}
