% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KellyCols.R
\docType{data}
\name{KellyCols}
\alias{KellyCols}
\title{Kelly colors of maximum contrast}
\format{An object of class \code{character} of length 22.}
\usage{
KellyCols
}
\value{
Vector containing names of Kelly colors
}
\description{
Kelly colors of maximum contrast
}
\keyword{datasets}
