% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndviCols.R
\docType{data}
\name{ndviCols}
\alias{ndviCols}
\title{Colors for NDVI plotting}
\format{
An object of class \code{character} of length 9.
}
\usage{
ndviCols
}
\value{
Vector containing names of Kelly colors
}
\description{
Colors for NDVI plotting
}
\keyword{datasets}
