% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pStars.R
\name{pStars}
\alias{pStars}
\title{Make a vector of asterisk strings for indicating p values}
\usage{
pStars(x, addSpace = FALSE)
}
\value{
vector
}
\description{
Make a vector of asterisk strings for indicating p values
}
