% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_pattern.R
\name{st_pattern}
\alias{st_pattern}
\title{Make a regular pattern MULTILINESTRING to "fill" an sf POLYGON or MULTIPOLYGON}
\usage{
st_pattern(x, size, pattern)
}
\value{
Polygon
}
\description{
Make a regular pattern MULTILINESTRING to "fill" an sf POLYGON or MULTIPOLYGON
}
