% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTheme.R
\docType{data}
\name{BasicTheme}
\alias{BasicTheme}
\title{A simple ggplot map theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 97.
}
\usage{
BasicTheme
}
\value{
Simple ggplot theme
}
\description{
A simple ggplot map theme
}
\keyword{datasets}
