% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptyTheme.R
\docType{data}
\name{EmptyTheme}
\alias{EmptyTheme}
\title{A ggplot theme that will make all background elements transparent}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 6.
}
\usage{
EmptyTheme
}
\value{
Simple ggplot theme
}
\description{
A ggplot theme that will make all background elements transparent
}
\keyword{datasets}
