% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showCols.R
\name{showCols}
\alias{showCols}
\title{Add a simple ggplot theme}
\usage{
showCols(cl = colors(), bg = "transparent", cex = 3, rot = 20)
}
\value{
Plot of colors in vector
}
\description{
Add a simple ggplot theme
}
